requirements.txt
Flask==2.2.5
flask_cors==4.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==2.2.2
PyPDF2==3.0.1
python-dotenv==1.0.0
openai==0.27.6
langchain==0.0.184
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0